Fix: Tailwind 적용 안됨 문제 해결 Skeleton 폴더 이동 후 Tailwind 적용 안되는 문제 해결
