+++ src/app/page.tsx
@@ -1,20 +1,5 @@
-//page.tsx
-'use client';
-
-import SK_Home from './skeleton/SK_Home';
-import MusicGrid from '@/components/MusicGrid';
-import useFirstMusicData from '@/hook/useFirstMusicData';
+import RootPage from '@/components/RootPage';
 
 export default function Home() {
-  const {status, firstMusicData} = useFirstMusicData();
-
-  if (status === 'pending') {
-    return <SK_Home />;
-  }
-
-  return (
-    <main className='flex-1 bg-music-background'>
-      <MusicGrid firstMusicData={firstMusicData} />
-    </main>
-  );
+  return <RootPage />;
 }
