+++ src/skeleton/SK_Home.tsx
@@ -4,7 +4,7 @@ import HomeTitle from '@/components/HomeTitle';
 
 export default function SK_Home() {
   return (
-    <main className='flex-1  bg-music-background'>
+    <main className='flex-1 bg-music-background'>
       <div className='container py-20'>
         <HomeTitle />
         <div className=' grid animate-pulse grid-cols-1 justify-between gap-x-20 gap-y-10 lg:grid-cols-2 xl:grid-cols-3'>
