+++ src/skeleton/SK_MusicCard.tsx
